+++ activation/e2e/nipost_test.go
@@ -180,7 +180,7 @@ func TestNIPostBuilderWithClients(t *testing.T) {
 	nb, err := activation.NewNIPostBuilder(
 		poetDb,
 		svc,
-		[]string{poetProver.RestURL().String()},
+		[]activation.PoetServer{{Address: poetProver.RestURL().String()}},
 		t.TempDir(),
 		logger.Named("nipostBuilder"),
 		sig,
@@ -233,7 +233,7 @@ func TestNIPostBuilder_Close(t *testing.T) {
 	nb, err := activation.NewNIPostBuilder(
 		poetDb,
 		svc,
-		[]string{poetProver.RestURL().String()},
+		[]activation.PoetServer{{Address: poetProver.RestURL().String()}},
 		t.TempDir(),
 		logger.Named("nipostBuilder"),
 		sig,
@@ -301,7 +301,7 @@ func TestNewNIPostBuilderNotInitialized(t *testing.T) {
 	nb, err := activation.NewNIPostBuilder(
 		poetDb,
 		svc,
-		[]string{poetProver.RestURL().String()},
+		[]activation.PoetServer{{Address: poetProver.RestURL().String()}},
 		t.TempDir(),
 		logger.Named("nipostBuilder"),
 		sig,