+++ activation/poet.go
@@ -18,6 +18,7 @@ import (
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/spacemeshos/go-spacemesh/common/types"
+	"github.com/spacemeshos/go-spacemesh/config/util"
 )
 
 var (
@@ -25,6 +26,11 @@ var (
 	ErrUnavailable    = errors.New("unavailable")
 	ErrInvalidRequest = errors.New("invalid request")
 )
+
+type PoetServer struct {
+	Address string         `mapstructure:"address"`
+	Pubkey  util.Base64Enc `mapstructure:"pubkey"`
+}
 
 type PoetPowParams struct {
 	Challenge  []byte
@@ -96,7 +102,7 @@ func WithLogger(logger *zap.Logger) PoetClientOpts {
 }
 
 // NewHTTPPoetClient returns new instance of HTTPPoetClient connecting to the specified url.
-func NewHTTPPoetClient(baseUrl string, cfg PoetConfig, opts ...PoetClientOpts) (*HTTPPoetClient, error) {
+func NewHTTPPoetClient(server PoetServer, cfg PoetConfig, opts ...PoetClientOpts) (*HTTPPoetClient, error) {
 	client := &retryablehttp.Client{
 		RetryMax:     cfg.MaxRequestRetries,
 		RetryWaitMin: cfg.RequestRetryDelay,
@@ -105,7 +111,7 @@ func NewHTTPPoetClient(baseUrl string, cfg PoetConfig, opts ...PoetClientOpts) (
 		CheckRetry:   checkRetry,
 	}
 
-	baseURL, err := url.Parse(baseUrl)
+	baseURL, err := url.Parse(server.Address)
 	if err != nil {
 		return nil, fmt.Errorf("parsing address: %w", err)
 	}
@@ -117,6 +123,9 @@ func NewHTTPPoetClient(baseUrl string, cfg PoetConfig, opts ...PoetClientOpts) (
 		baseURL: baseURL,
 		client:  client,
 		logger:  zap.NewNop(),
+		poetServiceID: types.PoetServiceID{
+			ServiceID: server.Pubkey.Bytes(),
+		},
 	}
 	for _, opt := range opts {
 		opt(poetClient)
@@ -183,18 +192,8 @@ func (c *HTTPPoetClient) Submit(
 }
 
 // PoetServiceID returns the public key of the PoET proving service.
-func (c *HTTPPoetClient) PoetServiceID(ctx context.Context) (types.PoetServiceID, error) {
-	if c.poetServiceID.ServiceID != nil {
-		return c.poetServiceID, nil
-	}
-	resBody := rpcapi.InfoResponse{}
-
-	if err := c.req(ctx, http.MethodGet, "/v1/info", nil, &resBody); err != nil {
-		return types.PoetServiceID{}, fmt.Errorf("getting poet ID: %w", err)
-	}
-
-	c.poetServiceID.ServiceID = resBody.ServicePubkey
-	return c.poetServiceID, nil
+func (c *HTTPPoetClient) PoetServiceID(ctx context.Context) types.PoetServiceID {
+	return c.poetServiceID
 }
 
 // Proof implements PoetProvingServiceClient.