+++ activation/poet_client_test.go
@@ -13,13 +13,14 @@ import (
 	"google.golang.org/protobuf/encoding/protojson"
 
 	"github.com/spacemeshos/go-spacemesh/common/types"
+	"github.com/spacemeshos/go-spacemesh/config/util"
 )
 
 func Test_HTTPPoetClient_ParsesURL(t *testing.T) {
 	cfg := server.DefaultRoundConfig()
 
 	t.Run("add http if missing", func(t *testing.T) {
-		client, err := NewHTTPPoetClient("bla", PoetConfig{
+		client, err := NewHTTPPoetClient(PoetServer{Address: "bla"}, PoetConfig{
 			PhaseShift: cfg.PhaseShift,
 			CycleGap:   cfg.CycleGap,
 		})
@@ -28,7 +29,7 @@ func Test_HTTPPoetClient_ParsesURL(t *testing.T) {
 	})
 
 	t.Run("do not change scheme if present", func(t *testing.T) {
-		client, err := NewHTTPPoetClient("https://bla", PoetConfig{
+		client, err := NewHTTPPoetClient(PoetServer{Address: "https://bla"}, PoetConfig{
 			PhaseShift: cfg.PhaseShift,
 			CycleGap:   cfg.CycleGap,
 		})
@@ -51,7 +52,7 @@ func Test_HTTPPoetClient_Submit(t *testing.T) {
 	defer ts.Close()
 
 	cfg := server.DefaultRoundConfig()
-	client, err := NewHTTPPoetClient(ts.URL, PoetConfig{
+	client, err := NewHTTPPoetClient(PoetServer{Address: ts.URL}, PoetConfig{
 		PhaseShift: cfg.PhaseShift,
 		CycleGap:   cfg.CycleGap,
 	}, withCustomHttpClient(ts.Client()))
@@ -83,7 +84,7 @@ func Test_HTTPPoetClient_Address(t *testing.T) {
 	defer ts.Close()
 
 	cfg := server.DefaultRoundConfig()
-	client, err := NewHTTPPoetClient(ts.URL, PoetConfig{
+	client, err := NewHTTPPoetClient(PoetServer{Address: ts.URL}, PoetConfig{
 		PhaseShift: cfg.PhaseShift,
 		CycleGap:   cfg.CycleGap,
 	}, withCustomHttpClient(ts.Client()))
@@ -105,7 +106,7 @@ func Test_HTTPPoetClient_Address_Mainnet(t *testing.T) {
 
 	for _, url := range poETServers {
 		t.Run(url, func(t *testing.T) {
-			client, err := NewHTTPPoetClient(url, PoetConfig{
+			client, err := NewHTTPPoetClient(PoetServer{Address: url}, PoetConfig{
 				PhaseShift: poetCfg.PhaseShift,
 				CycleGap:   poetCfg.CycleGap,
 			})
@@ -129,7 +130,7 @@ func Test_HTTPPoetClient_Proof(t *testing.T) {
 	defer ts.Close()
 
 	cfg := server.DefaultRoundConfig()
-	client, err := NewHTTPPoetClient(ts.URL, PoetConfig{
+	client, err := NewHTTPPoetClient(PoetServer{Address: ts.URL}, PoetConfig{
 		PhaseShift: cfg.PhaseShift,
 		CycleGap:   cfg.CycleGap,
 	}, withCustomHttpClient(ts.Client()))
@@ -153,12 +154,12 @@ func Test_HTTPPoetClient_PoetServiceID(t *testing.T) {
 	defer ts.Close()
 
 	cfg := server.DefaultRoundConfig()
-	client, err := NewHTTPPoetClient(ts.URL, PoetConfig{
+	key := util.MustBase64FromString("Zm9vYmFy")
+	client, err := NewHTTPPoetClient(PoetServer{Address: ts.URL, Pubkey: key}, PoetConfig{
 		PhaseShift: cfg.PhaseShift,
 		CycleGap:   cfg.CycleGap,
 	}, withCustomHttpClient(ts.Client()))
 	require.NoError(t, err)
 
-	_, err = client.PoetServiceID(context.Background())
-	require.NoError(t, err)
+	require.Equal(t, key.Bytes(), client.PoetServiceID(context.Background()).ServiceID)
 }