+++ cmd/root.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/spf13/cobra"
@@ -47,8 +48,6 @@ func AddCommands(cmd *cobra.Command) {
 		cfg.PublicMetrics.MetricsURL, "Push metrics to url")
 	cmd.PersistentFlags().DurationVar(&cfg.PublicMetrics.MetricsPushPeriod, "metrics-push-period",
 		cfg.PublicMetrics.MetricsPushPeriod, "Push period")
-	cmd.PersistentFlags().StringArrayVar(&cfg.PoETServers, "poet-server",
-		cfg.PoETServers, "The poet server url. (temporary) Can be passed multiple times")
 	cmd.PersistentFlags().StringVar(&cfg.Genesis.GenesisTime, "genesis-time",
 		cfg.Genesis.GenesisTime, "Time of the genesis layer in 2019-13-02T17:02:00+00:00 format")
 	cmd.PersistentFlags().StringVar(&cfg.Genesis.ExtraData, "genesis-extra-data",
@@ -295,6 +294,12 @@ func AddCommands(cmd *cobra.Command) {
 	cmd.PersistentFlags().StringVar(&cfg.TestConfig.SmesherKey, "testing-smesher-key",
 		"", "import private smesher key for testing",
 	)
+
+	/**========================  Deprecated flags ========================== **/
+	cmd.PersistentFlags().Var(flags.NewDeprecatedFlag(errors.New(
+		`The poet-server field is deprecated. Please migrate to the poet-servers field. `+
+			`Check 'Upgrade Information' in CHANGELOG.md for details.`,
+	)), "poet-server", "deprecated, use poet-servers instead")
 
 	// Bind Flags to config
 	err := viper.BindPFlags(cmd.PersistentFlags())